Added new aliases for pnpm and volta
In this update, we've introduced a few more handy shortcuts to improve your workflow. For pnpm, a new alias has been added to facilitate development tasks. On the volta side, we've expanded the node version management with additional aliases for installing and pinning Node.js version 22. Also, an alias was added for installing pnpm via Volta.