Added new aliases and updated PATH
New aliases have been added for dotnet, pnpm, and volta. The dotnet tool update command now has an alias. A new alias for the pnpm dev command was also introduced. For volta, additional node versions were included in the list of installable and pin-able versions. Furthermore, a new path to the dotnet tools directory was appended to the system PATH variable.